+++ metratec_rfid/deskid_uhf.py
@@ -21,7 +21,8 @@ class DeskIdUhf(UhfReaderAscii):
                 identification within the software and can be anything,
                 even an empty string.
 
-            serial_port (str): The serial port to use.
+            serial_port (str): The serial port to use. Leave empty (`""`)
+                to determine automatically during `connect()`.
 
         """
         super().__init__(instance, SerialConnection(serial_port))
@@ -40,7 +41,8 @@ class DeskIdUhfv2Base(UhfReaderAT, ReaderATSound):
                 identification within the software and can be anything,
                 even an empty string.
 
-            serial_port (str): The serial port to use.
+            serial_port (str): The serial port to use. Leave empty (`""`)
+                to determine automatically during `connect()`.
 
         """
         super().__init__(instance, SerialConnection(serial_port))