+++ metratec_rfid/qr_nfc.py
@@ -19,7 +19,8 @@ class QrNfc(NfcReaderAT):
                 identification within the software and can be anything,
                 even an empty string.
 
-            serial_port (str): The serial port to use.
+            serial_port (str): The serial port to use. Leave empty (`""`)
+                to determine automatically during `connect()`.
 
         """
         super().__init__(instance, SerialConnection(serial_port))