+++ metratec_rfid/quasar_lr.py
@@ -28,7 +28,8 @@ class QuasarLR(HfReaderAscii):
 
             port (int): The TCP connection port of the reader, defaults to 10001.
 
-            serial_port (str): The serial port of the reader.
+            serial_port (str): The serial port to use. Leave empty (`""`)
+                to determine automatically during `connect()`.
 
         """
         if hostname == "" and serial_port == "":